+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/stickers/StickerSet.kt
@@ -1,6 +1,7 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.stickers
 
 import com.github.insanusmokrassar.TelegramBotAPI.types.*
+import com.github.insanusmokrassar.TelegramBotAPI.types.files.PhotoSize
 import com.github.insanusmokrassar.TelegramBotAPI.types.files.Sticker
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
@@ -16,5 +17,7 @@ data class StickerSet(
     @SerialName(isAnimatedField)
     val isAnimated: Boolean = false,
     @SerialName(containsMasksField)
-    val containsMasks: Boolean = false
+    val containsMasks: Boolean = false,
+    @SerialName(thumbField)
+    val thumb: PhotoSize? = null
 )